+++ src/components/dialog_edit_entity.py
@@ -1,10 +1,10 @@
 from typing import List
 import streamlit as st
-import lib.state as state
 from lib.sparql_queries import find_entities, get_ontology
 from lib.sparql_base import insert, delete
-from schema import Entity
-from schema import Triple, DisplayTriple
+import lib.state as state
+from schema import Entity, Triple, DisplayTriple
+
 
 def __edit_entity(entity: Entity, triples_to_create: List[Triple], triples_to_delete: List[Triple]) -> None:
 
@@ -23,7 +23,6 @@ def __edit_entity(entity: Entity, triples_to_create: List[Triple], triples_to_de
     # Finalization: validation message and load the created entity
     state.set_toast('Entity Updated', ':material/done:')
     st.switch_page("pages/my-entities.py")
-
 
 
 @st.dialog('Edit an entity', width='large')
@@ -204,7 +203,7 @@ def dialog_edit_entity(entity: Entity, triples: List[DisplayTriple]) -> None:
 
 
     # Validation, edition, and display
-    if st.button('Edit', icon=":material/edit:"):
+    if st.button('Save', icon=":material/save:"):
         entity = Entity(
             uri=entity.uri, 
             label=new_entity_label.strip() if new_entity_label.strip() != '' else entity.label, # This might have changed